+++ extensions/ql-vscode/src/remote-queries/remote-query-result.ts
@@ -1,8 +1,10 @@
 import { DownloadLink } from './download-link';
+import { AnalysisFailure } from './shared/analysis-failure';
 
 export interface RemoteQueryResult {
   executionEndTime: Date;
   analysisSummaries: AnalysisSummary[];
+  analysisFailures: AnalysisFailure[];
 }
 
 export interface AnalysisSummary {